Renamed Rviz tool files
